:dragon_face: IMPLEMENTING lateral specific details part 1
