+++ src/pages/map_monitor_location/lateralDetailPanel.js
@@ -1,0 +1,52 @@
+// ** React Imports
+import { useState, useEffect }  from 'react';
+
+// **  Material Components Imports
+import { Box, Button, Typography, Menu, MenuItem,
+        Drawer, ToggleButtonGroup, ToggleButton, ListItemIcon
+      } from '@mui/material';
+
+// ** Icons Imports
+import ChevronDown from 'mdi-material-ui/ChevronDown'
+
+const LateralDeatilPanel = props => {
+
+  // * Props and states
+  const { controller } = props;
+  const [state, setState] = useState({ right: false });
+
+
+  const toogleDrawer = (anchor, open) => event => {
+    if(event.type === 'keydown' && (event.key === 'Tab' || event.key === 'Shift')) {
+      return;
+    }
+    setState ({...state, [anchor]: open});
+  }
+
+  return (
+    controller.id != null ? (
+      <Box>
+        <Button onClick={toogleDrawer('right',true)}>
+          {controller.controller_name}
+        </Button>
+        <Drawer
+          anchor={'right'}
+          open={state['right']}
+          onClose={toogleDrawer('right', false)}
+        >
+          <Box 
+            sx={{
+              backgroundColor: 'rgba(241,244,249,1)',
+              width: '100%',
+              height: '100%'
+            }}
+          >
+            {controller.local_area_controller_number+'번 '+controller.controller_name}
+          </Box>
+        </Drawer>
+      </Box>
+    ):('Here')
+  );
+}
+
+export { LateralDeatilPanel };
